Workplace from Facebook Template Spec (#5037)
* Workplace from Facebook Template Spec

* preview fix

* 3p correction

* 3p fix

* review fixes
